SimpleDataPool::Return resets data before returning back to pool
