+++ src/brpc/data_factory.h
@@ -24,19 +24,24 @@
 
 namespace brpc {
 
+// ---- thread safety ----
+// Method implementations of this interface should be thread-safe
 class DataFactory {
 public:
     virtual ~DataFactory() {}
 
-    // Implement this method to create a piece of data.
-    // Notice that this method is const.
+    // Implement this method to create a piece of data
     // Returns the data, NULL on error.
     virtual void* CreateData() const = 0;
 
-    // Implement this method to destroy a piece of data that was created
-    // by Create().
-    // Notice that this method is const.
+    // Implement this method to destroy data created by Create().
     virtual void DestroyData(void*) const = 0;
+
+    // Overwrite this method to reset the data before reuse. Nothing done by default.
+    // Returns
+    //   true:  the data can be kept for future reuse
+    //   false: the data was already destroyed and should NOT be reused.
+    virtual bool ResetData(void*) const { return true; }
 };
 
 } // namespace brpc
